[Chore] Decouple account and bucketlist api views
